requirements.txt
matplotlib==3.1.1
pyglet==1.3.2
gym[atari]==0.15.3
gym-retro==0.7.0
box2d==2.3.2
# because of Python 3.6
dataclasses==0.7
isort==4.3.21
black==19.10b0 
pytest==5.2.2
pytorch-ignite==0.2.1
xvfbwrapper==0.2.9
cma==2.7.0
pyflakes==2.1.1
